% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateChange.R
\name{climate_bar_plot}
\alias{climate_bar_plot}
\title{Simple Bar Chart of Global Temperature Anomalies}
\usage{
climate_bar_plot(filename)
}
\value{
A bar plot.
}
\description{
This function creates a bar chart of global temperature anomalies.
}
\note{
The plot is formatted for data sets found in Inst > Data > Extdata > ExcelFormattedGISTEMPData2CSV.csv
}
\examples{
\dontrun{

# Use the dataframe name from the import routine.

library(readr)
df<-import_gisstemp("ExcelFormattedGISTEMPData2CSV.csv")
library(ggplot2)
climate_bar_plot(df) 

}   

}
\references{
Current data sets can be found at \url{https://data.giss.nasa.gov/gistemp/}
}
