% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateChange.R
\name{climate_line_plot}
\alias{climate_line_plot}
\title{Simple Line Chart with Loess Smoother Global Temperature Anomalies}
\usage{
climate_line_plot(filename)
}
\value{
An line plot with a loess smoother.
}
\description{
This function creates a simple line chart with a loess smoother of global temperature anomalies.
}
\note{
The plot is formatted for data sets found in Inst > Data > Extdata > ExcelFormattedGISTEMPData2CSV.csv
}
\examples{
\dontrun{

# Use the dataframe name from the import routine.

library(readr)
df<-import_gisstemp("ExcelFormattedGISTEMPData2CSV.csv")
library(ggplot2)
climate_line_plot(df) 

}   

}
\references{
Current data sets can be found at \url{https://data.giss.nasa.gov/gistemp/}
}
