% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decade_increase.R
\name{decade_increase}
\alias{decade_increase}
\title{Inter-decade Change in Emissions Plot}
\usage{
decade_increase(df, decades, regions)
}
\arguments{
\item{df}{A dataframe containing annual carbon dioxide, methane and nitrous
oxide emissions records by country or region. It must include variables
\code{iso_code}, \code{year}, \code{co2}, \code{methane} and
\code{nitrous_oxide}, where \code{iso_code} is a code uniquely identifying
each region. The greenhouse gases emissions units should be in million
tonnes of carbon dioxide-equivalents. One such dataframe can obtained from
\emph{Our World in Data}: \url{https://github.com/owid/co2-data}.}

\item{decades}{A character vector containing the decades that should be
included in the plot, in the format \code{"1990-1999"}.}

\item{regions}{A character vector containing the codes of regions that should
be included in the plot. The codes must match those of the \code{df}
variable \code{iso_code}. If not provided or set to NULL, all of the
regions in \code{df} are included in the plot.}
}
\description{
\code{decade_increase} draws a plot of decade-to-decade change in total
greenhouse gas emissions (carbon dioxide, methane and nitrous oxide) for the
regions and decades provided as arguments. It also prints a table containing
average emissions for each country throughout the years that span the
decades provided in \code{decades} argument.
}
\examples{

Greenhouse_Gas_Emissions <-
read.csv("https://raw.githubusercontent.com/owid/co2-data/master/owid-co2-data.csv")

decade_increase(Greenhouse_Gas_Emissions,
c("1990-1999", "1980-1989"),
c("USA", "GBR"))

}
