% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions_pred.R
\name{emissions_pred}
\alias{emissions_pred}
\title{Prediction of Future Emissions Plots}
\usage{
emissions_pred(df, region, future)
}
\arguments{
\item{df}{A dataframe containing annual carbon dioxide, methane and nitrous
oxide emissions records by country or region, at least since 1990. It must
include variables \code{iso_code}, \code{year}, \code{co2}, \code{methane},
\code{nitrous_oxide}, \code{population} and \code{gdp} where
\code{iso_code} is a code uniquely identifying each region. The greenhouse
gases emissions units should be in million tonnes of carbon
dioxide-equivalents. One such dataframe can obtained from \emph{Our World
in Data}: \url{https://github.com/owid/co2-data}.}

\item{region}{A country or region code that matches one of the
\code{iso_code}s of \code{df}.}

\item{future}{A dataframe containing projected GDP and population data for
the region provided as argument to \code{region}. It must contain the
variables \code{year}, \code{population} and \code{gdp}. One such dataframe
for US, \code{futdata}, is included in the package.}
}
\description{
\code{lproj} draws plots with actual past and projected future annual
emissions from 1990. Three plots are provided in total, one for each of these
greenhouse gases: carbon dioxide, methane and nitrous oxide. The prediction
is based on the annual GDP, population and decade of the country or region of
interest.
}
\details{
The range of the actual past emissions curve goes from 1990 until the last
year of the dataframe provided to \code{df} argument. The projected emissions
curve ranges from the earliest to the latest year in the \code{future}
dataframe.
}
\examples{

data(futdata)
Greenhouse_Gas_Emissions <-
read.csv("https://raw.githubusercontent.com/owid/co2-data/master/owid-co2-data.csv")
emissions_pred(Greenhouse_Gas_Emissions, "USA", futdata)

}
