% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_data_f.R
\name{ex_data}
\alias{ex_data}
\title{ex_data}
\usage{
ex_data(big, small)
}
\arguments{
\item{big}{main data}

\item{small}{what you want to extract by}
}
\value{

}
\description{
the inputs should be raster or vectors
}
