% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_b_f.R
\name{get_a}
\alias{get_a}
\title{a}
\usage{
get_a(w, r, p, b)
}
\arguments{
\item{w}{quantile 60 percent of future data}

\item{r}{quantile 60 percent of control data}

\item{p}{g1 parameter}

\item{b}{b parameter}
}
\value{

}
\description{
transition parameter
}
