% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_b_f.R
\name{get_b}
\alias{get_b}
\title{b}
\usage{
get_b(g, qf, p, w, c, r)
}
\arguments{
\item{g}{g2 parameter}

\item{qf}{quantile 90 percent of future data}

\item{p}{g1 parameter}

\item{w}{quantile 60 percent of future data}

\item{c}{quantile 90 percent of control data}

\item{r}{quantile 60 percent of control data}
}
\value{

}
\description{
transition parameter
}
