% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_e_f.R
\name{get_e}
\alias{get_e}
\title{get_e}
\usage{
get_e(pr, qn)
}
\arguments{
\item{pr}{input data that should be higher than 90 percent quantile}

\item{qn}{quantile 90 percent of the input}
}
\value{

}
\description{
the function find you the mean ex, the input data needs to be higher than 90 percent quantile
}
