% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_g1_g2_f.R
\name{get_g1}
\alias{get_g1}
\title{get_g1}
\usage{
get_g1(oq, cq)
}
\arguments{
\item{oq}{quantile 60 percent of observed data}

\item{cq}{quantile 60 percent of control data}
}
\value{

}
\description{
get_g1
}
