% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_g1_g2_f.R
\name{get_g2}
\alias{get_g2}
\title{get_g2}
\usage{
get_g2(oq, cq)
}
\arguments{
\item{oq}{quantile 90 percent of observed data}

\item{cq}{quantile 90 percent of control data}
}
\value{

}
\description{
get_g2
}
