% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tem_trans_f.R
\name{get_tem_trans}
\alias{get_tem_trans}
\title{get_tem_trans}
\usage{
get_tem_trans(mo, mf, mc, to, sd)
}
\arguments{
\item{mo}{mean of observed T}

\item{mf}{mean of upcoming T}

\item{mc}{mean of control T}

\item{to}{observed T}

\item{sd}{difference of standard deviation sdt_f/sdt_c}
}
\value{

}
\description{
it finds the transition of T
}
