% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateChange.R
\name{import_gisstemp}
\alias{import_gisstemp}
\title{Import GISSTEMP data}
\usage{
import_gisstemp(filename)
}
\arguments{
\item{filename}{Filename in csv format to be read. The file must be located in the working directory.}
}
\value{
Returns a data frame from an Excel file into a data frame.
}
\description{
This function imports a file formatted CSV file.
}
\details{
Current data sets for modeling climate change can be found at the following address in csv formats:
\url{https://data.giss.nasa.gov/gistemp/}
}
\examples{
# Download ExcelFormattedGISTEMPData2CSV.csv file to your working directory to play with the data.
# Returns a dataframe with the default name "gisstemp" which can be used for line or bar plots.
\dontrun{

library(readr)
df<-import_gisstemp("ExcelFormattedGISTEMPData2CSV.csv")

}
}
