% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lproj.R
\name{lproj}
\alias{lproj}
\title{Linear Prediction of Past Emissions Plots}
\usage{
lproj(df, region)
}
\arguments{
\item{df}{A dataframe containing annual carbon dioxide, methane and nitrous
oxide emissions records by country or region, at least since 1990. It must
include variables \code{iso_code}, \code{year}, \code{co2}, \code{methane},
\code{nitrous_oxide}, \code{population} and \code{gdp} where
\code{iso_code} is a code uniquely identifying each region. The greenhouse
gases emissions units should be in million tonnes of carbon
dioxide-equivalents. One such dataframe can obtained from \emph{Our World
in Data}: \url{https://github.com/owid/co2-data}.}

\item{region}{A country or region code that matches one of the
\code{iso_code}s of \code{df}.}
}
\description{
\code{lproj} draws plots with actual and predicted annual emissions since
1990 to the latest year in the dataframe provided in \code{df} argument.
Three plots are provided in total, one for each of these greenhouse gases:
carbon dioxide, methane and nitrous oxide. The prediction is based on the
annual GDP, population and decade of the country or region of interest.
}
\examples{

Greenhouse_Gas_Emissions <-
read.csv("https://raw.githubusercontent.com/owid/co2-data/master/owid-co2-data.csv")
lproj(df = Greenhouse_Gas_Emissions, region = "USA")

}
