% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsData.R
\docType{data}
\name{obs}
\alias{obs}
\title{obs data}
\format{
data table with 3 columns
\describe{
  \item{DTM}{recording time}
  \item{pr_o}{observed precipitation}
  \item{tem_o}{observed tempreture}
}
}
\usage{
obs
}
\description{
observed precipitation and temperature data
}
\keyword{datasets}
