% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_emissions.R
\name{plot_emissions}
\alias{plot_emissions}
\title{Plot of annual greenhouse gas emissions}
\usage{
plot_emissions(df, regions = NULL, title = NULL)
}
\arguments{
\item{df}{A dataframe containing annual carbon dioxide, methane and nitrous
oxide emissions records by country or region. It must include variables
\code{iso_code}, \code{year}, \code{co2}, \code{methane} and
\code{nitrous_oxide}, where \code{iso_code} is a code uniquely identifying
each region. The greenhouse gases emissions units should be in million
tonnes of carbon dioxide-equivalents. One such dataframe can obtained from
\emph{Our World in Data}: \url{https://github.com/owid/co2-data}.}

\item{regions}{A character vector containing the codes of regions that should
be included in the plot. The codes must match those of the \code{df}
variable \code{iso_code}. If not provided or set to NULL, all of the
regions in \code{df} are included in the plot.}

\item{title}{The title of the plot. It is NULL by default.}
}
\description{
\code{plot_emissions} draws a plot of annual greenhouse gas emissions for the
countries provided to the \code{regions} argument.
}
\details{
The x-axis range of the plot goes from the earliest to the latest year with
available data for any of the regions provided to the \code{regions}
argument. If the number of regions requested is between two and ten, a legend
is included to the plot.
}
\examples{
Greenhouse_Gas_Emissions <-
read.csv("https://raw.githubusercontent.com/owid/co2-data/master/owid-co2-data.csv")

plot_emissions(Greenhouse_Gas_Emissions,
c("USA", "CHN", "GBR", "RUS"),
"Emissions of historical global powers")

}
