% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_me_a_b_f.R
\name{plot_me_a_b}
\alias{plot_me_a_b}
\title{plot_me_a_b}
\usage{
plot_me_a_b(a, b)
}
\arguments{
\item{a}{a input "transition parameter"}

\item{b}{b input "transition parameter"}
}
\value{

}
\description{
visualizing a and b parameters
}
