% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_gcmData.R
\docType{data}
\name{pr_gcm}
\alias{pr_gcm}
\title{pr_gcm}
\format{
RasterBrick
\describe{
  \item{dimensions}{16, 17, 272, 55115  (nrow, ncol, ncell, nlayers)}
  \item{resolution}{2.8125, 2.790426  (x, y)}
  \item{extent}{-1.40625, 46.40625, 27.90615, 72.55296  (xmin, xmax, ymin, ymax)}
  \item{crs}{+proj=longlat +datum=WGS84 +no_defs }
  \item{Date}{1950-01-01, 2100-12-31 (min, max)}
   \item{varname}{pr}






}
}
\usage{
pr_gcm
}
\description{
simulated precipitation
}
\keyword{datasets}
