% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_trans_less_f.R
\name{pr_trans_less}
\alias{pr_trans_less}
\title{pr_trans_less}
\usage{
pr_trans_less(a, pr, b)
}
\arguments{
\item{a}{transition parameter}

\item{pr}{observed precipitation}

\item{b}{transition parameter}
}
\value{

}
\description{
finding the transition precipitation for the value less than quantile 90 percent
}
