% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_trans_more_f.R
\name{pr_trans_more}
\alias{pr_trans_more}
\title{pr_trans_more}
\usage{
pr_trans_more(E, pr, q, a, b)
}
\arguments{
\item{E}{the difference of e_f/e_c}

\item{pr}{observed precipitation}

\item{q}{quantile 90 percent of observed data}

\item{a}{transition parameter a}

\item{b}{transition parameter b}
}
\value{

}
\description{
finding the transition precipitation for the value more than quantile 90 percent
}
