% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateChange.R
\name{stream_graph_non_zero}
\alias{stream_graph_non_zero}
\title{Stream graph with a non zero offset}
\usage{
stream_graph_non_zero(filename)
}
\arguments{
\item{filename}{Filename is a dataframe that was imported for region2.xlsx}
}
\value{
Returns a streamgraph interpretation of a bar plot.
}
\description{
This function creates an interactive streamgraph with a non zero offset.
}
\details{
This function uses library(streamgraph).
}
\examples{
# Download region2.xlsx file to your working directory to play with the data.
# Returns a dataframe with the default name "region2" which can be used for an interactive streamgraph.
\dontrun{

library(streamgraph)
library(readxl)
df<-import_gisstemp_streamgraph("region2.xlsx")
stream_graph_non_zero(df)

}
}
